+++ make/mcu/STM32F4.mk
@@ -208,6 +208,7 @@ VCP_SRC = \
 endif
 
 MSC_SRC = \
+            drivers/usb_msc_common.c \
             drivers/usb_msc_f4xx.c \
             msc/usbd_msc_desc.c \
             msc/usbd_storage.c
