+++ make/mcu/STM32H7.mk
@@ -256,6 +256,7 @@ MCU_EXCLUDES = \
             drivers/timer.c
 
 #MSC_SRC = \
+#            drivers/usb_msc_common.c \
 #            drivers/usb_msc_h7xx.c \
 #            msc/usbd_storage.c
