+++ src/main/drivers/usb_msc_f7xx.c
@@ -37,40 +37,21 @@
 
 #include "blackbox/blackbox.h"
 #include "drivers/io.h"
-#include "drivers/light_led.h"
 #include "drivers/nvic.h"
-#include "drivers/persistent.h"
 #include "drivers/serial_usb_vcp.h"
 #include "drivers/system.h"
 #include "drivers/time.h"
 #include "drivers/usb_msc.h"
 
-#include "drivers/accgyro/accgyro_mpu.h"
+#include "msc/usbd_storage.h"
 
 #include "pg/sdcard.h"
 #include "pg/usb.h"
 
 #include "vcp_hal/usbd_cdc_interface.h"
+
 #include "usb_io.h"
 #include "usbd_msc.h"
-#include "msc/usbd_storage.h"
-
-#define DEBOUNCE_TIME_MS 20
-
-static IO_t mscButton;
-
-void mscInit(void)
-{
-    if (usbDevConfig()->mscButtonPin) {
-        mscButton = IOGetByTag(usbDevConfig()->mscButtonPin);
-        IOInit(mscButton, OWNER_USB_MSC_PIN, 0);
-        if (usbDevConfig()->mscButtonUsePullup) {
-            IOConfigGPIO(mscButton, IOCFG_IPU);
-        } else {
-            IOConfigGPIO(mscButton, IOCFG_IPD);
-        }
-    }
-}
 
 uint8_t mscStart(void)
 {
@@ -124,65 +105,6 @@ uint8_t mscStart(void)
     NVIC_EnableIRQ(SysTick_IRQn);
 
     return 0;
-}
-
-bool mscCheckBoot(void)
-{
-    const uint32_t bootModeRequest = persistentObjectRead(PERSISTENT_OBJECT_RESET_REASON);
-    return bootModeRequest == RESET_MSC_REQUEST;
-    // Note that we can't clear the persisent object after checking here. This is because
-    // this function is called multiple times during initialization. So we clear on a reset
-    // out of MSC mode.
-}
-
-bool mscCheckButton(void)
-{
-    bool result = false;
-    if (mscButton) {
-        uint8_t state = IORead(mscButton);
-        if (usbDevConfig()->mscButtonUsePullup) {
-            result = state == 0;
-        } else {
-            result = state == 1;
-        }
-    }
-
-    return result;
-}
-
-void mscWaitForButton(void)
-{
-    // In order to exit MSC mode simply disconnect the board, or push the button again.
-    while (mscCheckButton());
-    delay(DEBOUNCE_TIME_MS);
-    while (true) {
-        asm("NOP");
-        if (mscCheckButton()) {
-            systemResetFromMsc();
-        }
-    }
-}
-
-void systemResetToMsc(int timezoneOffsetMinutes)
-{
-    persistentObjectWrite(PERSISTENT_OBJECT_RESET_REASON, RESET_MSC_REQUEST);
-
-    __disable_irq();
-
-    // Persist the RTC across the reboot to use as the file timestamp
-#ifdef USE_PERSISTENT_MSC_RTC
-    rtcPersistWrite(timezoneOffsetMinutes);
-#else
-    UNUSED(timezoneOffsetMinutes);
-#endif
-    NVIC_SystemReset();
-}
-
-void systemResetFromMsc(void)
-{
-    persistentObjectWrite(PERSISTENT_OBJECT_RESET_REASON, RESET_NONE);
-    __disable_irq();
-    NVIC_SystemReset();
 }
 
 #endif
